+++ src/main/java/fr/pilato/elasticsearch/tools/template/TemplateReader.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package fr.pilato.elasticsearch.tools.componenttemplate;
+package fr.pilato.elasticsearch.tools.template;
 
 import fr.pilato.elasticsearch.tools.SettingsFinder.Defaults;
 import fr.pilato.elasticsearch.tools.SettingsReader;
@@ -27,36 +27,38 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 
 /**
- * Manage elasticsearch index template files in _index_templates dir
- * and component template files in _component_templates dir.
+ * Manage elasticsearch template files
  * @author David Pilato
  */
-public class ComponentTemplateSettingsReader extends SettingsReader {
+@Deprecated
+public class TemplateReader extends SettingsReader {
 
-	private static final Logger logger = LoggerFactory.getLogger(ComponentTemplateSettingsReader.class);
+	private static final Logger logger = LoggerFactory.getLogger(TemplateReader.class);
 
 	/**
-	 * Read a component template
+	 * Read a legacy template
 	 * @param root dir within the classpath
 	 * @param template template name (.json will be appended)
-	 * @return The component template content
-	 * @throws IOException if we can not read the file
+	 * @return The template content
+	 * @throws IOException if the connection with elasticsearch is failing
 	 */
-	public static String readComponentTemplate(String root, String template) throws IOException {
+	@Deprecated
+	public static String readTemplate(String root, String template) throws IOException {
 		if (root == null) {
-			return readComponentTemplate(template);
+			return readTemplate(template);
 		}
-		String settingsFile = root + "/" + Defaults.ComponentTemplatesDir + "/" + template + Defaults.JsonFileExtension;
+		String settingsFile = root + "/" + Defaults.TemplateDir + "/" + template + Defaults.JsonFileExtension;
 		return readFileFromClasspath(settingsFile);
 	}
 
 	/**
-	 * Read a component template in default classpath dir
-	 * @param componentTemplate component template name (.json will be appended)
-	 * @return The component template content
-	 * @throws IOException if we can not read the file
+	 * Read a legacy template in default classpath dir
+	 * @param template template name (.json will be appended)
+	 * @return The template content
+	 * @throws IOException if the connection with elasticsearch is failing
 	 */
-	public static String readComponentTemplate(String componentTemplate) throws IOException {
-		return readComponentTemplate(Defaults.ConfigDir, componentTemplate);
+	@Deprecated
+	public static String readTemplate(String template) throws IOException {
+		return readTemplate(Defaults.ConfigDir, template);
 	}
 }
